requirements.txt
requests==2.25.1
Django==3.2
django-environ==0.4.5
django-configurations==2.2
django-waffle==2.1.0
deal==4.6.0
pytest-django==4.2.0
wheel==0.36.2
celery==5.1.2

pydocstyle==5.1.1
flake8==3.8.4
flake8-2020==1.6.0
flake8-blind-except==0.2.0
flake8-bugbear==20.11.1
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.3.1
flake8-debugger==4.0.0
flake8-docstrings==1.5.0
flake8-eradicate==1.0.0
flake8-polyfill==1.0.2
flake8-print==4.0.0
flake8-quotes==3.2.0
flake8-string-format==0.3.0
flake8-fixme==1.1.1
flake8-annotations-complexity==0.0.6
flake8-variables-names==0.0.4
flake8-class-attributes-order==0.1.2
flake8-broken-line==0.3.0
flake8-tidy-imports==4.2.1
flake8-typing-imports==1.10.1
dlint==0.11.0
flake8-if-statements==0.1.0
flake8-functions==0.0.5
flake8-annotations-coverage==0.0.5
flake8-expression-complexity==0.0.9
flake8-printf-formatting==1.1.2
flake8-multiline-containers==0.0.17
flake8-absolute-import==1.0
flake8-simplify==0.13.0
flake8-functions-names==0.0.5
